requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.7.7
django-environ==0.4.5
idna==2.8
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
PyMySQL==1.0.2